requirements.txt
dash==0.38.0
dash-html-components==0.13.5
dash-core-components==0.43.1
python-dotenv==0.10.1
PyGithub==1.43.5
pytest==4.3.0